CountryListForm クラスの page, size フィールドの型を long → int へ修正しました。

+++ src/main/java/ksbysample/webapp/basic/web/CountryListForm.java
@@ -20,8 +20,8 @@ public class CountryListForm {
     @Size(max = 45, message = "{error.size.max}")
     private String localName;
 
-    private long page;
+    private int page;
 
-    private long size;
+    private int size;
 
 }
